+++ src/io_loop.rs
@@ -5,6 +5,7 @@ use crate::{
     frames::Frames,
     heartbeat::Heartbeat,
     internal_rpc::InternalRPCHandle,
+    killswitch::KillSwitch,
     protocol::{self, AMQPError, AMQPHardError},
     socket_state::{SocketEvent, SocketState},
     thread::ThreadHandle,
@@ -44,6 +45,7 @@ pub struct IoLoop {
     connection_io_loop_handle: ThreadHandle,
     stream: Pin<Box<dyn AsyncIOHandle + Send>>,
     status: Status,
+    killswitch: KillSwitch,
     frame_size: FrameSize,
     receive_buffer: Buffer,
     send_buffer: Buffer,
@@ -66,6 +68,7 @@ impl IoLoop {
             protocol::constants::FRAME_MIN_SIZE,
             configuration.frame_max(),
         );
+        let killswitch = heartbeat.killswitch();
 
         Ok(Self {
             connection_status,
@@ -78,6 +81,7 @@ impl IoLoop {
             connection_io_loop_handle,
             stream,
             status: Status::Initial,
+            killswitch,
             frame_size,
             receive_buffer: Buffer::with_capacity(FRAMES_STORAGE * frame_size as usize),
             send_buffer: Buffer::with_capacity(FRAMES_STORAGE * frame_size as usize),
@@ -159,18 +163,33 @@ impl IoLoop {
                     let mut readable_context = Context::from_waker(&readable_waker);
                     let writable_waker = self.writable_waker();
                     let mut writable_context = Context::from_waker(&writable_waker);
+                    let mut res = Ok(());
                     while self.should_continue() {
                         if let Err(err) = self.run(&mut readable_context, &mut writable_context) {
-                            self.critical_error(err)?;
+                            res = self.critical_error(err);
                         }
                     }
-                    self.internal_rpc.stop();
                     self.heartbeat.cancel();
-                    Ok(())
+                    let internal_rpc = self.internal_rpc.clone();
+                    if self.killswitch.killed() {
+                        internal_rpc.register_internal_future(std::future::poll_fn(move |cx| {
+                            self.stream
+                                .as_mut()
+                                .poll_close(cx)
+                                .map(|res| res.map_err(From::from))
+                        }));
+                    }
+                    internal_rpc.stop();
+                    res
                 })?,
         );
         waker.wake();
         Ok(())
+    }
+
+    fn stop(&mut self) {
+        self.status = Status::Stop;
+        self.heartbeat.cancel();
     }
 
     fn poll_socket_events(&mut self) {
@@ -179,7 +198,7 @@ impl IoLoop {
 
     fn check_connection_state(&mut self) {
         if self.connection_status.closed() {
-            self.status = Status::Stop;
+            self.stop();
         }
     }
 
@@ -226,7 +245,7 @@ impl IoLoop {
         if let Some(resolver) = self.connection_status.connection_resolver() {
             resolver.swear(Err(error.clone()));
         }
-        self.status = Status::Stop;
+        self.stop();
         self.channels.set_connection_error(error.clone());
         for (_, resolver) in std::mem::take(&mut self.serialized_frames) {
             if let Some(resolver) = resolver {
@@ -334,6 +353,8 @@ impl IoLoop {
 
                 if let Some(sz) = self.socket_state.handle_read_poll(res) {
                     if sz > 0 {
+                        self.heartbeat.update_last_read();
+
                         trace!("read {} bytes", sz);
                         self.receive_buffer.fill(sz);
                     } else {